+++ webrtc/modules/audio_coding/main/acm2/acm_receiver_unittest.cc
@@ -115,12 +115,12 @@ class AcmReceiverTest : public AudioPacketizationCallback,
     }
   }
 
-  virtual int SendData(
+  virtual int32_t SendData(
       FrameType frame_type,
       uint8_t payload_type,
       uint32_t timestamp,
       const uint8_t* payload_data,
-      uint16_t payload_len_bytes,
+      size_t payload_len_bytes,
       const RTPFragmentationHeader* fragmentation) OVERRIDE {
     if (frame_type == kFrameEmpty)
       return 0;
