+++ webrtc/modules/audio_coding/main/acm2/audio_coding_module_unittest_oldapi.cc
@@ -87,7 +87,7 @@ class PacketizationCallbackStub : public AudioPacketizationCallback {
       uint8_t payload_type,
       uint32_t timestamp,
       const uint8_t* payload_data,
-      uint16_t payload_len_bytes,
+      size_t payload_len_bytes,
       const RTPFragmentationHeader* fragmentation) OVERRIDE {
     CriticalSectionScoped lock(crit_sect_.get());
     ++num_calls_;
