+++ webrtc/modules/audio_coding/main/test/TestAllCodecs.h
@@ -29,12 +29,14 @@ class TestPack : public AudioPacketizationCallback {
   void RegisterReceiverACM(AudioCodingModule* acm);
 
   virtual int32_t SendData(
-      FrameType frame_type, uint8_t payload_type,
-      uint32_t timestamp, const uint8_t* payload_data,
-      uint16_t payload_size,
+      FrameType frame_type,
+      uint8_t payload_type,
+      uint32_t timestamp,
+      const uint8_t* payload_data,
+      size_t payload_size,
       const RTPFragmentationHeader* fragmentation) OVERRIDE;
 
-  uint16_t payload_size();
+  size_t payload_size();
   uint32_t timestamp_diff();
   void reset_payload_size();
 
@@ -45,7 +47,7 @@ class TestPack : public AudioPacketizationCallback {
   uint32_t timestamp_diff_;
   uint32_t last_in_timestamp_;
   uint64_t total_bytes_;
-  uint16_t payload_size_;
+  size_t payload_size_;
 };
 
 class TestAllCodecs : public ACMTest {
@@ -61,7 +63,7 @@ class TestAllCodecs : public ACMTest {
   // This is useful for codecs which support several sampling frequency.
   // Note! Only mono mode is tested in this test.
   void RegisterSendCodec(char side, char* codec_name, int32_t sampling_freq_hz,
-                         int rate, int packet_size, int extra_byte);
+                         int rate, int packet_size, size_t extra_byte);
 
   void Run(TestPack* channel);
   void OpenOutFile(int test_number);
@@ -75,7 +77,7 @@ class TestAllCodecs : public ACMTest {
   PCMFile outfile_b_;
   int test_count_;
   int packet_size_samples_;
-  int packet_size_bytes_;
+  size_t packet_size_bytes_;
 };
 
 }  // namespace webrtc
